Alarm Trigger added
I added an alarm trigger to the the program, to make it I used time picker which is not compatable with strings. this is an easy fix however it does mess with the database and toggling times. for now they both work, but but together. We also need to make it so the alarm stops when the picture is taken. Other then that we are pretty solid with our orginal idea and we can start moving foward with the image recognition and the back log.